fix: time column at table / add value after pipe
