+++ package.json
@@ -24,6 +24,7 @@
     "@types/node": "^17.0.35",
     "@types/react": "^17.0.45",
     "@types/react-dom": "^18.0.5",
+    "@types/react-router-dom": "^5.3.3",
     "axios": "^0.25.0",
     "classnames": "^2.3.1",
     "date-fns": "^2.28.0",
@@ -49,8 +50,7 @@
     "slate": "^0.73.1",
     "slate-history": "^0.66.0",
     "slate-react": "^0.74.2",
-    "tinycolor2": "^1.4.2",
-    "typescript": "^4.6.4"
+    "tinycolor2": "^1.4.2"
   },
   "scripts": {
     "webpack": "webpack",
@@ -115,6 +115,8 @@
     "sass-loader": "^12.4.0",
     "style-loader": "^3.3.1",
     "terser-webpack-plugin": "^5.3.1",
+    "ts-loader": "^9.3.0",
+    "typescript": "^4.6.4",
     "url-loader": "^4.1.1",
     "webpack": "^5.67.0",
     "webpack-cli": "^4.9.2",
