+++ src/App.tsx
@@ -1,6 +1,7 @@
 import { Provider } from "react-redux";
 import MainView from "./components/MainView";
 import store from "./store/store";
+
 export default function App() {
     return (
         <Provider store={store}>
