+++ src/components/LabelBrowser/helpers/querybuilder.js
@@ -26,7 +26,8 @@ export function queryBuilder(labels, hasPipe = false, pipeLabels = []) {
 
     if (hasPipe) {
         postTags = actualQuery.split("}")[1];
-        const expParse = actualQuery.split(/[|]/);
+        const json = /[|json]/
+        const expParse = actualQuery.split(json);
         if (pipeParseOpts.some((s) => expParse[1].includes(s))) {
             const pipeTags = ` | ${pipeLabels}`;
             postTags = postTags.toString().concat(pipeTags.toString());
@@ -63,8 +64,9 @@ export function queryBuilder(labels, hasPipe = false, pipeLabels = []) {
 }
 
 export function queryBuilderWithLabels(hasPipe = false, pipeLabels = []) {
-    const labels = store.getState().labels;
 
+    const labels = store.getState().labels;
+console.log(labels)
     const query = queryBuilder(labels, hasPipe, pipeLabels);
 
     store.dispatch(setQuery(query));
